Combined profile.jsx into the uploadPicture.jsx
